wire up USB-Connected and Battery-Charging headers
